Fix the MinPropertiesRequiredTautology canonicalizer rule
