+++ DEPENDENCIES
@@ -1,5 +1,5 @@
 vendorpull https://github.com/sourcemeta/vendorpull 6c398482e30bd85d7e8138c606e01f28270c29ae
 rapidjson https://github.com/Tencent/rapidjson 06d58b9e848c650114556a23294d0b6440078c61
-jsontoolkit https://github.com/sourcemeta/jsontoolkit a0ba1a35a1708941fc81a4bb6125f9bd178f50a4
+jsontoolkit https://github.com/sourcemeta/jsontoolkit 85bbbc240242a343e9a6ab16cb2ae5ffdb9de1b6
 googletest https://github.com/google/googletest e2239ee6043f73722e7aa812a459f54a28552929
 bootstrap https://github.com/twbs/bootstrap 1a6fdfae6be09b09eaced8f0e442ca6f7680a61e
